BOL-375 refactor: api 서버 맞춰서 추가
