+++ data/src/main/java/com/yapp/bol/data/remote/FileApi.kt
@@ -9,6 +9,7 @@ import retrofit2.http.POST
 import retrofit2.http.Part
 
 interface FileApi {
+
     @Multipart
     @POST("v1/file")
     suspend fun postFileUpload(
