+++ data/src/main/java/com/yapp/bol/data/remote/GroupApi.kt
@@ -1,19 +1,13 @@
 package com.yapp.bol.data.remote
 
-import com.yapp.bol.data.model.base.BaseResponse
-import com.yapp.bol.data.model.group.GuestAddApiRequest
-import com.yapp.bol.data.model.group.JoinGroupApiRequest
-import com.yapp.bol.data.model.group.MemberListResponse
 import com.yapp.bol.data.model.group.request.CheckGroupJonByAccessCodeRequest
-import com.yapp.bol.data.model.group.response.GameApiResponse
-import com.yapp.bol.data.model.group.response.MemberValidApiResponse
 import com.yapp.bol.data.model.group.request.NewGroupApiRequest
 import com.yapp.bol.data.model.group.response.CheckGroupJoinByAccessCodeResponse
 import com.yapp.bol.data.model.group.response.GroupDetailResponse
-import com.yapp.bol.data.model.group.response.NewGroupApiResponse
 import com.yapp.bol.data.model.group.response.GroupSearchApiResponse
-import com.yapp.bol.data.model.group.response.JoinedGroupResponse
+import com.yapp.bol.data.model.group.response.NewGroupApiResponse
 import com.yapp.bol.data.model.group.response.RandomImageResponse
+import com.yapp.bol.data.model.rank.UserRankApiResponse
 import retrofit2.Response
 import retrofit2.http.Body
 import retrofit2.http.GET
@@ -28,16 +22,11 @@ interface GroupApi {
         @Body newGroupApiRequest: NewGroupApiRequest,
     ): Response<NewGroupApiResponse>
 
-    @GET("/v1/group/{groupId}/game")
-    suspend fun getGameList(
+    @GET("/v1/group/{groupId}/game/{gameId}")
+    suspend fun getUserRank(
         @Path("groupId") groupId: Int,
-    ): Response<GameApiResponse>
-
-    @GET("/v1/group/{groupId}/member/validateNickname")
-    suspend fun getValidateNickName(
-        @Path("groupId") groupId: Int,
-        @Query("nickname") nickName: String,
-    ): Response<MemberValidApiResponse>
+        @Path("gameId") gameId: Int,
+    ): Response<UserRankApiResponse>
 
     @GET("/v1/group/default-image")
     suspend fun getRandomImage(): Response<RandomImageResponse>
@@ -49,37 +38,14 @@ interface GroupApi {
         @Query("pageSize") pageSize: String,
     ): Response<GroupSearchApiResponse>
 
-    @GET("/v1/user/me/group")
-    suspend fun getJoinedGroup(): Response<JoinedGroupResponse>
-
     @GET("/v1/group/{groupId}")
     suspend fun getGroupDetail(
         @Path("groupId") groupId: Long
     ): Response<GroupDetailResponse>
-
-    @GET("/v1/group/{groupId}/member")
-    suspend fun getMemberList(
-        @Path("groupId") groupId: Int,
-        @Query("size") pageSize: Int,
-        @Query("cursor") cursor: String?,
-        @Query("nickname") nickname: String?,
-    ): Response<MemberListResponse>
-
-    @POST("/v1/group/{groupId}/guest")
-    suspend fun postGuestMember(
-        @Path("groupId") groupId: Int,
-        @Body guestAddApiRequest: GuestAddApiRequest
-    )
 
     @POST("v1/group/{groupId}/accessCode")
     suspend fun checkGroupJoinAccessCode(
         @Path("groupId") groupId: String,
         @Body accessCode: CheckGroupJonByAccessCodeRequest,
     ): Response<CheckGroupJoinByAccessCodeResponse>
-
-    @POST("v1/group/{groupId}/host")
-    suspend fun joinGroup(
-        @Path("groupId") groupId: String,
-        @Body request: JoinGroupApiRequest,
-    ): Response<BaseResponse>
 }
